+++ db/schemas/trusted/ddl_logistics_01.sql
@@ -7,6 +7,6 @@ CREATE TABLE IF NOT EXISTS td_ext_anp.logistics_01 (
     sub_classificacao STRING,
     operacao STRING,
     modal STRING,
-    qtd_produto_liquido FLOAT64,
+    qtd_produto_liquido NUMERIC,
     data_criacao TIMESTAMP DEFAULT CURRENT_TIMESTAMP()
 ) PARTITION BY DATE(data_criacao);
